% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{display_modality}
\alias{display_modality}
\title{Make a Plot of Modality}
\usage{
display_modality(fluster_obj)
}
\arguments{
\item{fluster_obj}{A fluster object.}
}
\description{
Positivity thresholds are calculated by finding the compact extremes
of marker expressions in various clusters.  Thresholds are the mid-point between
extremes.
}
