% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{examine_positivity_thresholds}
\alias{examine_positivity_thresholds}
\title{Examine Positivity Thresholds}
\usage{
examine_positivity_thresholds(
  fcs,
  parameters = NULL,
  skew.crit = 0.2,
  manual_thresholds = NULL
)
}
\arguments{
\item{fcs}{Either a flowSet or a flowFrame, as will be provided to fluster().}

\item{parameters}{A vector of parameters to be considered.}

\item{skew.crit}{A list of clusters we want to gate.}

\item{manual_thresholds}{A named vector of one or more values that will override
the internally-calculated positivity thresholds.}
}
\value{
A modality object that can be passed to fluster().
}
\description{
Given fcs data and a list of markers (parameters), calculate
and visualize positivity thresholds.
}
