% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster_add_mst}
\alias{fluster_add_mst}
\title{Add a minimum spanning tree representation to the fluster object}
\usage{
fluster_add_mst(fluster_obj)
}
\value{
A fluster object with the graph slot populated.
}
\description{
Add a minimum spanning tree representation to the fluster object
#param fluster_obj The result of running fluster()
}
