% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster_add_tsne}
\alias{fluster_add_tsne}
\title{Add a tSNE representation to the fluster object}
\usage{
fluster_add_tsne(fluster_obj)
}
\value{
A fluster object with the tsne slot populated.
}
\description{
Add a tSNE representation to the fluster object
#param fluster_obj The result of running fluster()
}
