% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster_gate_clusters}
\alias{fluster_gate_clusters}
\title{Gate Events in Cluster(s)}
\usage{
fluster_gate_clusters(fluster_obj, fcs_obj, clusters)
}
\arguments{
\item{fluster_obj}{An object of type "fluster", the result of running fluster()}

\item{fcs_obj}{A flowFrame or flowSet to be gated.}

\item{clusters}{A list of clusters we want to gate.}
}
\value{
The gated flowFrame or flowSet
}
\description{
Given a fluster model and a list of its clusters, gate a flowFrame
or flowSet so as to include events that below to these clusters.
}
