% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster_map_sample}
\alias{fluster_map_sample}
\title{Map a Sample to a Fluster Model}
\usage{
fluster_map_sample(ff, fluster_obj)
}
\arguments{
\item{ff}{A sample flowFrame}

\item{fluster_obj}{An object of type "fluster", the result of running fluster()}
}
\value{
Per-cluster counts and fractions
}
\description{
This function determines, for a single sample, the number of cells in each cluster.
}
