% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{map_functional_names}
\alias{map_functional_names}
\title{title map_functional_names}
\usage{
map_functional_names(fluster_obj, defs_file)
}
\arguments{
\item{fluster_obj}{The result of running fluster()}

\item{defs_file}{A file containing the functional definitions.}
}
\value{
A decorated fluster object
}
\description{
Based on a user-specified table, assign symbolic names to clusters
}
