% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{merge_categorical_clusters}
\alias{merge_categorical_clusters}
\title{Merge Categorically Similar Clusters}
\usage{
merge_categorical_clusters(
  fluster_obj,
  parameters = colnames(fluster_obj$centers),
  sd_fac = 1
)
}
\arguments{
\item{fluster_obj}{A fluster object.}

\item{parameters}{The parameters of the fluster object.}

\item{sd_fac}{A factor multiplying the standard deviation to determine if that
marker is sufficiently above (below) the threshold in order to labeled "hi" ("lo").}
}
\value{
A fluster object after categorical merging.
}
\description{
Clusters are labeled with a categorical vector in which each
marker is either "hi" or "lo" with respect to a threshold.  If a marker is not unambiguously
either hi or lo, it's labeled as "un" for "unknown.  To receive hi (lo), the
cluster center must be sufficiently above (below) the threshold in units of
the standard deviation of that marker.
}
