% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster}
\alias{fluster}
\title{Fingerprint-based Clustering}
\usage{
fluster(
  fcs,
  parameters = NULL,
  nRecursions = 12,
  nclust = NULL,
  merge = TRUE,
  graph = TRUE,
  tsne = TRUE,
  manual_thresholds = NULL,
  skew.crit = 0.2,
  modality = NULL,
  sd_fac = 1
)
}
\arguments{
\item{fcs}{The data (either a flowFrame or a flowSet)}

\item{parameters}{The parameters in fcs used for analysis}

\item{nRecursions}{The number of recursions in calculating the fingerprint (default = 12)}

\item{nclust}{The number of clusters you want panoply to make.  If NULL, fluster}

\item{merge}{Logical: should we merge initial clusters based on categorical similarity?
will make a guess as to the "best" number of clusters}

\item{graph}{Logical: should we compute the MST for visualization?}

\item{tsne}{Logical: should we compute a tSNE embedding for visualization?}

\item{manual_thresholds}{A named vector of one or more values that will override
the internally-calculated positivity thresholds.  If NULL (default) use the internally-calculated
thresholds.}

\item{skew.crit}{Numeric (default = 0.2).  for unimodally-distributed markers,
if abs(skewness) of the global distribution of a parameter is smaller than this
criterion, that parameter is deemed uninformative and will be dropped from further
consideration (unless overridden via manual_thresholds).  See Description.}

\item{modality}{If previously computed with examine_positivity_thresholds, you can pass it
in.  If NULL, modality is computed internally.}

\item{sd_fac}{A factor by which to multiply per-marker cluster standard deviation to determine
positivity.}
}
\value{
An object of class 'fluster', with the following elements:
\describe{
\item{mod}{The flowFPModel generated}
\item{centers}{Multivariate bin centers}
\item{graph}{A graph that can be used for visualization}
\item{clustering}{A named list containing cluster membership of the bins}
}
}
\description{
Fluster (\strong{F}ingerprint-based c\strong{luster}ing)
implements a workflow of doing Cytometric Fingerprint (CF) binning of
a flowFrame (or flowSet)
using flowFP to a relatively high resolution (default nRecursions of 12 results
in 4096 bins).  The bin centroids are then computed, using the median value of all
of the events in the bin for each of the included parameters.  Next, these
multivariate bin centroids are clustered using agglommerative hierarchical clustering
\code{cluster::agnes()}.  The resulting data are represented in a graph structure and or as a tSNE embedding
for visualization and interpretation.

The parameter skew.crit determines whether a marker is considered to be informative.
In order to force ALL markers to be considered, set skew.crit = 0.0.
}
\examples{
load(system.file("extdata", "sampled_flowset_young.rda", package = "fluster"))
flust_params = c(7:9, 11:22)
flust_obj = fluster(fs_young, parameters = flust_params)
}
