% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{fluster_phenobars}
\alias{fluster_phenobars}
\title{Visualize Cluster Phenotypes}
\usage{
fluster_phenobars(
  fluster_obj,
  parameters = colnames(fluster_obj$centers),
  cluster = 1,
  bin_indices = NULL,
  plot_global_flag = FALSE,
  show_thresholds = TRUE,
  show_sd_fac = TRUE,
  main = paste("Cluster", cluster)
)
}
\arguments{
\item{fluster_obj}{An object of type "fluster", the result of running fluster()}

\item{parameters}{Which parameters to include in the plot (default = all parameters)}

\item{cluster}{Which cluster to plot.}

\item{bin_indices}{Instead of bin indices in a cluster, specify them directly.}

\item{plot_global_flag}{Indicate the global distributions.}

\item{show_thresholds}{Logical. Show per-parameter thresholds.}

\item{show_sd_fac}{Logical.  Superimpose modified error flags if sd_fac != 1.0.}

\item{main}{Title of plot.}
}
\description{
Draw a "phenobar" representation of a cluster phenotype.  Bars have
a height equalt to the medial value of the parameter and are
color-coded.  Error flags represent first and third quartiles of the bin centers
belonging to the cluster.
}
