% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{plot_fluster_graph}
\alias{plot_fluster_graph}
\title{plot_fluster_graph}
\usage{
plot_fluster_graph(
  fluster,
  markers = colnames(fluster$mat),
  vs = 10,
  ms = 5,
  log.size = FALSE,
  vertex.frame = TRUE,
  cex.main = 2,
  cex.lab = 2
)
}
\arguments{
\item{fluster}{The result of running fluster}

\item{vs}{The max size of nodes in the graph}

\item{ms}{The minimum size of nodes in the graph}

\item{cex.main}{Scale factor for titles of the individual markers}

\item{log_size}{If true, scale node sizes logarithmically}

\item{vertex_frame}{Logical.  Should we draw a frame.}
}
\value{
N/A.
}
\description{
Draw a picture of the result of fluster using graph-based representation of clusters.
}
\examples{
plot_fluster(fluster_obj)
}
