% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluster_methods.R
\name{plot_fluster_tsne}
\alias{plot_fluster_tsne}
\title{plot_fluster_tsne}
\usage{
plot_fluster_tsne(
  fluster,
  markers = NULL,
  mode = c("arithmetic", "robust"),
  cex = 50,
  proportional = TRUE,
  emph = TRUE,
  cex.lab = 2,
  highlight_clusters = NULL,
  legend = TRUE
)
}
\arguments{
\item{fluster}{The result of running fluster}

\item{markers}{Markers to include in the spread}

\item{mode}{Compute colors using either arithmetic or robust average}

\item{cex}{Scale factor for node size}

\item{proportional}{Logical.  Scale by the number of events in the cluster}

\item{emph}{Logical.  Emphasize each blob with a black line.}

\item{cex.lab}{Scale factor for titles of the individual markers}

\item{highlight_clusters}{IF not NULL, a collection of cluster indices to highlight.}

\item{legend}{Logical.  Whether or not to draw a legend.}
}
\value{
N/A.
}
\description{
Draw a picture of the result of fluster using tsne representation of clusters.
}
\examples{
plot_fluster(fluster_obj)
}
